+++ RELEASE_NOTES.md
@@ -10,8 +10,9 @@ This version bumps the package to Swift 5.9 and adds support for visionOS.
 This version also bumps the deployment targets to make it possible to add more features. 
 
 ### ✨ Features
-swfi
+
 * `FlowLayout` is now public.
+* `TagCapsuleStyle` has new, material-based default styles.
 * `TagCapsuleStyle` now supports specifying a background material as well.
 
 ### 💡 Behavior Changes